+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changes in 0.6.18 ()
 Improvements:
  * Upgrade MatrixKit version ().
  * RoomVC: Add a re-request keys button on message unable to decrypt (#1879).
- * Analytics: Move code from AppDelegate to a dedicated class: Analytics
- 
+ * Analytics: Move code from AppDelegate to a dedicated class: Analytics.
+ * Analytics: Track Matrix SDK stats (time to startup the app).
+
 Bug fix:
 
 Changes in 0.6.17 (2018-06-01)